fix: improve show more button behaviour and affordance
