--- components-ui/see-more/style.module.css
@@ -1,21 +1,0 @@
-.expanded {
-}
-.container {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-.collapsed {
-  height: 0rem;
-  position: relative;
-  overflow: hidden;
-}
-.collapsed::after {
-  content: '';
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  height: 1rem;
-  background: linear-gradient(rgba(255, 255, 255, 0), #fff);
-}
