+++ components/espace-agent-components/documents/data-section/actes.tsx
@@ -4,7 +4,7 @@ import routes from '#clients/routes';
 import { Warning } from '#components-ui/alerts';
 import ButtonLink from '#components-ui/button';
 import { PrintNever } from '#components-ui/print-visibility';
-import SeeMore from '#components-ui/see-more';
+import ShowMore from '#components-ui/show-more';
 import { DataSectionClient } from '#components/section/data-section/client';
 import { FullTable } from '#components/table/full';
 import { EAdministration } from '#models/administrations/EAdministration';
@@ -58,11 +58,12 @@ export const AgentActesSection: React.FC<{
                 Cette entreprise possède {documents.actes.length} document(s) au
                 RNE. Chaque document peut contenir un ou plusieurs actes :
               </p>
-              <ActesTable actes={documents.actes.slice(0, 5)} />
-              {documents.actes.length > 5 && (
-                <SeeMore aria-label="Voir le reste des actes">
-                  <ActesTable actes={documents.actes.slice(5)} />
-                </SeeMore>
+              {documents.actes.length >= 5 ? (
+                <ShowMore>
+                  <ActesTable actes={documents.actes} />
+                </ShowMore>
+              ) : (
+                <ActesTable actes={documents.actes} />
               )}
             </>
           )
